Create GoodQuestion.txt
The more details the better. Include anything relevant to the problem, especially what you have tried and what has failed. In this case, the source code would be in the folder...at your discretion. Error messages are helpful too, even if they are vague. It's important to communicate the problem as precisely as possible and give others the basic pieces they need to work with.